distance function finished

+++ graph.cpp
@@ -267,7 +267,7 @@ Edge Graph::removeEdge(Vertex source, Vertex destination)
 }
 
 
-Edge Graph::setEdgeWeight(Vertex source, Vertex destination, int weight)
+Edge Graph::setEdgeWeight(Vertex source, Vertex destination, float weight)
 {
     if (assertEdgeExists(source, destination, __func__) == false)
         error("edge does not exist");
